requirements.txt
python-telegram-bot[job-queue]==20.7
supabase==2.3.0
python-dotenv==1.0.0
aiosqlite==0.19.0
httpx==0.27.2
